Use StringValueCStr rather than StringValuePtr
The Kerberos API expects C Strings (or rather, it doesn't accept a
length argument, so the only reasonable expectation are C Strings).
StringValuePtr will return a C String in most cases, but when a String
contains null-bytes, it cannot be represented as a C String. If this is
the case, StringValueCStr raises an ArgumentError and execution of the
current method is halted. Simply passing the StringValuePtr (which
should only be used in conjunction with RSTRING_LEN) leads to
potentially incorrect strings being sent to kerberos.